+++ mrb_doc/models/texture2d.rb
@@ -13,4 +13,17 @@ class Texture2D
     # src/mruby_integration/models/texture2d.cpp
     Texture2D.new
   end
+
+  # Unloads the {Texture2D} from memory.
+  #
+  # @example Basic usage
+  #   texture = Texture2D.new("/assets/my_cool_sprite.png")
+  #   texture.unload
+  #
+  # @return [nil]
+  def unload
+    # mrb_Texture2D_unload
+    # src/mruby_integration/models/texture2d.cpp
+    nil
+  end
 end