+++ CHANGELOG.md
@@ -253,6 +253,7 @@
 - Removed Texture2D#height=
 - Removed Texture2D#mipmaps=
 - Removed Texture2D#format=
+- Texture2D#unload is now C++ backed
 
 ## v0.3.14.1